+++ docs/changelog/changelog.rst
@@ -5,8 +5,9 @@ develop
 * new cli command noun `checkpoint` (in the next major release this tap will be deprecated
 * DataContext.create() adds an additional directory: `checkpoints`
 * DataContext.list_checkpoints() returns a list of checkpoint names found in the project
-* new cli command: checkpoint list
 * DataContext.get_checkpoint() returns a validated dictionary loaded from yml
+* new cli command: checkpoint list
+* new cli command: checkpoint run
 
 
 0.10.4